[appBuilder] 增加删除、修改知识库配置时，同组其他成员 is_default 的处理

+++ app-knowledge/services/knowledge-service/src/main/java/modelengine/jade/knowledge/KnowledgeCenterService.java
@@ -48,7 +48,7 @@ public interface KnowledgeCenterService {
      */
     @ToolMethod(name = "delete_user_knowledge_config", description = "删除用户的知识库配置信息")
     @Genericable(id = "knowledge.center.service.deleteUserKnowledgeConfig")
-    void delete(@Property(description = "知识库配置id", required = true) int id);
+    void delete(@Property(description = "知识库配置id", required = true) Long id);
 
     /**
      * 查询用户的知识库配置信息
